+++ example/lib/examples/dotlottie.dart
@@ -1,0 +1,39 @@
+import 'package:collection/collection.dart';
+import 'package:flutter/material.dart';
+import 'package:lottie/lottie.dart';
+
+void main() => runApp(const App());
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      home: Scaffold(
+        body: Example(),
+      ),
+    );
+  }
+}
+
+//---- example
+class Example extends StatelessWidget {
+  const Example({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Lottie.asset(
+      'assets/cat.lottie',
+      decoder: customDecoder,
+    );
+  }
+}
+
+Future<LottieComposition?> customDecoder(List<int> bytes) {
+  return LottieComposition.decodeZip(bytes, filePicker: (files) {
+    return files.firstWhereOrNull(
+        (f) => f.name.startsWith('animations/') && f.name.endsWith('.json'));
+  });
+}
+//----