+++ lib/src/animation/keyframe/gradient_color_keyframe_animation.dart
@@ -1,3 +1,4 @@
+import 'dart:math' as math;
 import 'dart:ui';
 import '../../model/content/gradient_color.dart';
 import '../../value/keyframe.dart';
@@ -8,8 +9,17 @@ class GradientColorKeyframeAnimation extends KeyframeAnimation<GradientColor> {
 
   GradientColorKeyframeAnimation(List<Keyframe<GradientColor>> keyframes)
       : super(keyframes) {
-    var startValue = keyframes.first.startValue;
-    var size = startValue == null ? 0 : startValue.size;
+    // Not all keyframes that this GradientColor are used for will have the same length.
+    // AnimatableGradientColorValue.ensureInterpolatableKeyframes may add extra positions
+    // for some keyframes but not others to ensure that it is interpolatable.
+    // Ensure that there is enough space for the largest keyframe.
+    var size = 0;
+    for (var i = 0; i < keyframes.length; i++) {
+      var startValue = keyframes[i].startValue;
+      if (startValue != null) {
+        size = math.max(size, startValue.size);
+      }
+    }
     _gradientColor = GradientColor(List<double>.filled(size, 0.0),
         List<Color>.filled(size, const Color(0x00000000)));
   }