+++ lib/src/providers/file_provider.dart
@@ -1,6 +1,5 @@
 import 'dart:ui' as ui;
-import 'package:flutter/foundation.dart';
-import 'package:path/path.dart' as p;
+import 'package:flutter/widgets.dart';
 import '../composition.dart';
 import '../lottie_image_asset.dart';
 import 'load_image.dart';
@@ -9,17 +8,20 @@ import 'provider_io.dart' if (dart.library.html) 'provider_web.dart' as io;
 
 @immutable
 class FileLottie extends LottieProvider {
-  FileLottie(this.file, {super.imageProviderFactory});
+  FileLottie(
+    this.file, {
+    super.imageProviderFactory,
+    super.decoder,
+  });
 
   final Object /*io.File|html.File*/ file;
 
   @override
-  Future<LottieComposition> load() {
+  Future<LottieComposition> load({BuildContext? context}) {
     return sharedLottieCache.putIfAbsent(this, () async {
       var bytes = await io.loadFile(file);
-      var composition = await LottieComposition.fromBytes(bytes,
-          name: p.basenameWithoutExtension(io.filePath(file)),
-          imageProviderFactory: imageProviderFactory);
+      var composition =
+          await LottieComposition.fromBytes(bytes, decoder: decoder);
 
       for (var image in composition.images.values) {
         image.loadedImage ??= await _loadImage(composition, image);