+++ lib/src/providers/lottie_provider.dart
@@ -4,9 +4,11 @@ import '../../lottie.dart';
 import 'load_image.dart';
 
 abstract class LottieProvider {
-  LottieProvider({this.imageProviderFactory});
+  LottieProvider({this.imageProviderFactory, this.decoder});
 
   final LottieImageProviderFactory? imageProviderFactory;
+
+  final LottieDecoder? decoder;
 
   ImageProvider? getImageProvider(LottieImageAsset lottieImage) {
     var imageProvider = fromDataUri(lottieImage.fileName);
@@ -16,7 +18,7 @@ abstract class LottieProvider {
     return imageProvider;
   }
 
-  Future<LottieComposition> load();
+  Future<LottieComposition> load({BuildContext? context});
 }
 
 class LottieCache {